refactored to app router
